add useParent option
Added a useParent option for when we don't want the text to overflow the parent box. Basically, we get the height of the parent and subtract the height of all its children (excluding the element itself) to calculate the height of the box. As an example, we may have a card with an image and text, we don't want the text to overflow the card but the image may take up a different height each time.